FE Refactoring splitted task form into separate modules, task type enum
