+++ myexaminer-frontend/src/components/exam/CreateExam.js
@@ -7,7 +7,7 @@ import Tabs from '@material-ui/core/Tabs';
 import Tab from '@material-ui/core/Tab';
 import Typography from '@material-ui/core/Typography';
 import Box from '@material-ui/core/Box';
-import TaskForm from "./TaskForm";
+import TaskForm from "./task-form/TaskForm";
 
 function TabPanel(props) {
   const { children, value, index, ...other } = props;
