+++ myexaminer-frontend/src/components/exam/task-form/TaskTypeEnum.js
@@ -1,0 +1,7 @@
+const TaskTypeEnum = Object.freeze({
+  BLANKS: "L",
+  CLOSED: "Z",
+  OPEN: "O",
+});
+
+export default TaskTypeEnum;
